doc snap

+++ doc/3d-snapshot/Base.c
@@ -19,11 +19,11 @@ BaseValidateUlong(
 )
 {
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - Pos0);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - Pos0);
   uint64_t positionAfterUlong;
   if (hasBytes)
   {
-    positionAfterUlong = Pos0 + (uint64_t)(uint32_t)4U;
+    positionAfterUlong = Pos0 + (uint64_t)4U;
   }
   else
   {
