+++ doc/3d-snapshot/BoundedSum.c
@@ -25,11 +25,11 @@ ValidateBoundedSumLeft(
 {
   /* SNIPPET_START: boundedSum */
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAfterBoundedSum;
   if (hasBytes)
   {
-    positionAfterBoundedSum = Pos + (uint64_t)(uint32_t)4U;
+    positionAfterBoundedSum = Pos + (uint64_t)4U;
   }
   else
   {
@@ -75,11 +75,11 @@ ValidateBoundedSumRight(
 {
   /* Validating field right */
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAfterBoundedSum;
   if (hasBytes)
   {
-    positionAfterBoundedSum = Pos + (uint64_t)(uint32_t)4U;
+    positionAfterBoundedSum = Pos + (uint64_t)4U;
   }
   else
   {
@@ -221,11 +221,11 @@ ValidateMySumBound(
 {
   /* Validating field bound */
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAftermySum;
   if (hasBytes)
   {
-    positionAftermySum = Pos + (uint64_t)(uint32_t)4U;
+    positionAftermySum = Pos + (uint64_t)4U;
   }
   else
   {
