+++ doc/3d-snapshot/EnumConstraint.c
@@ -41,11 +41,11 @@ ValidateEnumConstraintX(
 {
   /* Validating field x */
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAfterEnumConstraint;
   if (hasBytes)
   {
-    positionAfterEnumConstraint = Pos + (uint64_t)(uint32_t)4U;
+    positionAfterEnumConstraint = Pos + (uint64_t)4U;
   }
   else
   {
@@ -119,11 +119,11 @@ EnumConstraintValidateEnumConstraint(
 {
   /* Validating field col */
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - StartPosition);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - StartPosition);
   uint64_t positionAfterEnumConstraint;
   if (hasBytes)
   {
-    positionAfterEnumConstraint = StartPosition + (uint64_t)(uint32_t)4U;
+    positionAfterEnumConstraint = StartPosition + (uint64_t)4U;
   }
   else
   {
