+++ doc/3d-snapshot/ReadPair.c
@@ -26,11 +26,11 @@ ValidatePairFirst(
 {
   /* Validating field first */
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAfterPair;
   if (hasBytes)
   {
-    positionAfterPair = Pos + (uint64_t)(uint32_t)4U;
+    positionAfterPair = Pos + (uint64_t)4U;
   }
   else
   {
@@ -113,11 +113,11 @@ ValidatePairSecond(
 {
   /* Validating field second */
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)4U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)4U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAfterPair;
   if (hasBytes)
   {
-    positionAfterPair = Pos + (uint64_t)(uint32_t)4U;
+    positionAfterPair = Pos + (uint64_t)4U;
   }
   else
   {
