+++ doc/3d-snapshot/Triangle.c
@@ -25,11 +25,11 @@ ValidatePointX(
 {
   /* Validating field x */
   /* Checking that we have enough space for a UINT16, i.e., 2 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)2U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)2U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAfterPoint;
   if (hasBytes)
   {
-    positionAfterPoint = Pos + (uint64_t)(uint32_t)2U;
+    positionAfterPoint = Pos + (uint64_t)2U;
   }
   else
   {
@@ -68,11 +68,11 @@ ValidatePointY(
 {
   /* Validating field y */
   /* Checking that we have enough space for a UINT16, i.e., 2 bytes */
-  BOOLEAN hasBytes = (uint64_t)(uint32_t)2U <= ((uint64_t)Input.len - Pos);
+  BOOLEAN hasBytes = (uint64_t)2U <= ((uint64_t)Input.len - Pos);
   uint64_t positionAfterPoint;
   if (hasBytes)
   {
-    positionAfterPoint = Pos + (uint64_t)(uint32_t)2U;
+    positionAfterPoint = Pos + (uint64_t)2U;
   }
   else
   {
